Versión 1.1 Refactoring de aplicacion
To-do
Cambiar el acceso a los datos.
