+++ src/logica/RestriccionesDominio.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ag_pi1;
+package logica;
 
 /**
  *
